Realizzata prima versione dello scale in spazio tridimensionale. CameraTargetMovement interferisce sulla posizione di target che deve restare immutata rispetto metodi esterni, da rivedere la logica della camera

+++ Game2Dprj/CustomEventArgs.cs
@@ -20,11 +20,11 @@ namespace Game2Dprj
 
     public class TrackerGameEventArgs : EventArgs
     {
-        public TrackerGameEventArgs(int accuracy)
+        public TrackerGameEventArgs(double accuracy)
         {
             Accuracy = accuracy;
         }
 
-        public int Accuracy { get; set; }
+        public double Accuracy { get; set; }
     }
 }
